Fixed tests for UI hooks
