+++ enthought_example/example_ui_hooks/tests/test_ui_notification_hooks_factory.py
@@ -2,12 +2,11 @@ import unittest
 
 from envisage.api import Plugin
 
-from force_wfmanager.plugins.ui_notification.ui_notification_hooks_factory \
-    import \
-    UINotificationHooksFactory
-from force_wfmanager.plugins.ui_notification.ui_notification_hooks_manager \
-    import \
-    UINotificationHooksManager
+from enthought_example.example_ui_hooks.example_ui_hooks_factory import (
+    ExampleUIHooksFactory)
+from enthought_example.example_ui_hooks.example_ui_hooks_manager import (
+    ExampleUIHooksManager
+)
 
 try:
     import mock
@@ -15,15 +14,15 @@ except ImportError:
     from unittest import mock
 
 
-class TestUINotificationHooksFactory(unittest.TestCase):
+class TestExampleUIHooksFactory(unittest.TestCase):
     def test_initialization(self):
         mock_plugin = mock.Mock(spec=Plugin)
-        factory = UINotificationHooksFactory(mock_plugin)
+        factory = ExampleUIHooksFactory(mock_plugin)
         self.assertEqual(factory.plugin, mock_plugin)
 
     def test_create_ui_hooks_manager(self):
         mock_plugin = mock.Mock(spec=Plugin)
-        factory = UINotificationHooksFactory(mock_plugin)
+        factory = ExampleUIHooksFactory(mock_plugin)
         self.assertIsInstance(
             factory.create_ui_hooks_manager(),
-            UINotificationHooksManager)
+            ExampleUIHooksManager)
